requirements.txt
altgraph==0.17.5
asgiref==3.11.0
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
pygame==2.6.1
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pywin32-ctypes==0.2.3
reportlab==4.4.5
setuptools==80.9.0
sqlparse==0.5.4
tzdata==2025.2
